changed script tag to pageend.ejs added images
