--- course/week-1.md
@@ -1,54 +1,0 @@
-# Exercises Week 1
-
-Minor Web Development - Progressive Web Apps
-
-## Intention
-
-This week we will convert the client side OBA app to a server side version. We use Node.js & Express for this. We will also use npm as a build tool and use modules and Browserify to perform optimization for performance.
-
-## Method
-
-12 hours have been scheduled to work independently on this week's assignments this week. Try to plan your work wisely! In the meantime, progress is monitored in standup meetings and classroom meetings. At the end of the week you will be tested on theoretical knowledge and on the understanding of the code that you write according to the assignments.
-
-## Exercises
-
-1. Refactor the OBA or WAFS App to a server side version with Node.js & Express
-2. Tooling
-3. Universal / Isomorphic app (extra)
-4. Static Site Generator / Serverless (extra)
-
-### Exercise 1: Refactor the OBA or WAFS App to a server side version with Node.js & Express
-
-When refactoring, think carefully about the MVP (or core functionality) of the OBA app. In general this is an overview page, with a list of items that each refer to a detail page. Make sure that you first generate the HTML of the server, then you will see to what extent you can adjust the CSS and finally you try to improve the app as much as possible with client-side JavaScript.
-
-Fork this repo for your exercises: [Progressive Web Apps 1920](https://github.com/cmda-minor-web/progressive-web-apps-1920)
-
-Document in your readme.md how you will get the server up and running; git clone && npm start
-
-#### Resources
-
-* [Introduction to Node.js](https://egghead.io/courses/introduction-to-node-the-fundamentals)
-* [Getting started with Express.js](https://egghead.io/courses/getting-started-with-express-js)
-* [Export modules with Node.js](https://egghead.io/lessons/node-js-export-modules-in-node-js)
-
-### Exercise 2: Tooling
-
-Install and implement the tooling for your app. You can choose for example a workflow with NPM scripts, modules and Browserify.
-
-Document in your readme.md how you have implemented the build
-
-#### Resources
-
-[Introduction to NPM scripts](https://medium.freecodecamp.org/introduction-to-npm-scripts-1dbb2ae01633)
-
-### Exercise 3: Universal / Isomorphic app (extra)
-
-With the arrival of Node.js it is possible to run JavaScript on the server. A Universal or Isomorphic app is an app whose (part of) the code can be used on both the client and the server. Try to turn your app into a universal app.
-
-#### Resources
-
-[https://github.com/wooorm/dictionary](https://github.com/wooorm/dictionary)
-
-### Exercise 4: Jam stack; Static Site Generators / Serverless (extra)
-
-Node.js; seen that, done that, bought a t-shirt? Combine a static site generator with a serverless implementation.
