--- course/week-2.md
@@ -1,34 +1,0 @@
-# Opdrachten Week 2
-
-Minor Web Development - Progressive Web Apps
-
-## Intention
-
-This week we convert our OBA or WAFS application to a Progressive Web App
-
-## Method
-
-12 hours have been scheduled to work independently on this week's assignments this week. Try to plan your work wisely! In the meantime, progress is monitored in standup meetings and classroom meetings. At the end of the week you will be tested on theoretical knowledge and on the understanding of the code that you write according to the assignments.
-
-## Exercises
-
-1. Convert your app into a Progressive Web App
-2. Implement a Service Worker
-
-### Exercise 1: Convert your app into a Progressive Web App
-
-Make an installable version of your app, making it feel more app-like.
-
-#### Resources
-
-...
-
-### Exercise 2: Implement a Service Worker
-
-Consider how you want to implement the Service Worker in your OBA app. Create a job story for this. Make sure that you at least cache and serve a number of static assets with the Worker. Also provide feedback to the user about the online / offline state.
-
-Include your job story in the README.md
-
-#### Resources
-
-...
