--- course/week-3.md
@@ -1,43 +1,0 @@
-# Exercises Week 3
-
-Minor Web Development - Progressive Web apps
-
-## Intention
-
-Deze week gaan we de critical rendering path van de OBA applicatie optimaliseren.
-
-## Method
-
-16 hours have been scheduled to work independently on this week's assignments this week. Try to plan your work wisely! In the meantime, progress is monitored in standup meetings and classroom meetings. At the end of the week you will be tested on theoretical knowledge and on the understanding of the code that you write according to the assignments.
-
-## Exercises
-
-1. Optimize the performance (Critical Render Path) of your application
-2. Install your app on an online web server
-
-### Exercise 1: Optimize the performance (Critical Render Path) of your application
-
-Optimize 2 or more of the performance components below, also explain in the README.md why you thought it was important to choose these performance optimizations:
-
-* First view
-* Repeat view
-* Perceived performance
-* Image loading
-* Runtime performance
-* Time to first byte
-
-Make sure you work out every optimization in a separate branch. Merge the branches back to the master branch.
-Document in your README.md, per optimization, what the effect of the optimization has been. Ultimately you also give a conclusion about the total effect of your optimizations.
-
-#### Resources
-
-1. [Why performance matters](https://developers.google.com/web/fundamentals/performance/why-performance-matters/)
-2. [Understanding the GitHub flow](https://guides.github.com/introduction/flow/)
-
-### Exercise 2: Install your app on an online web server
-
-Install your app on an online web server. Then the web app can not only be viewed locally but also online.
-
-#### Resources
-
-[Digital Ocean](https://www.digitalocean.com/)
