+++ docs/static/js/script.js
@@ -1,13 +1,14 @@
 console.log('test')
 
-if ('serviceWorker' in navigator) {
-    window.addEventListener('load', function() {
-      navigator.serviceWorker.register('/service/sw.js').then(function(registration) {
-        // Registration was successful
-        console.log('ServiceWorker registration successful with scope: ', registration.scope);
-      }, function(err) {
-        // registration failed :(
-        console.log('ServiceWorker registration failed: ', err);
-      });
-    });
-  }
+// if ('serviceWorker' in navigator) {
+//     window.addEventListener('load', function() {
+//       navigator.serviceWorker.register('/service/sw.js').then(function(registration) {
+//         // Registration was successful
+//         console.log('ServiceWorker registration successful with scope: ', registration.scope);
+//         return registration.update();
+//       }, function(err) {
+//         // registration failed :(
+//         console.log('ServiceWorker registration failed: ', err);
+//       });
+//     });
+//   }
