+++ docs/static/manifest.webmanifest
@@ -4,17 +4,47 @@
     "description": "Dota2 progress application",
     "icons": [
       {
-        "src": "/img/dota-2-logo-192.png",
-        "type": "image/png",
-        "sizes": "192x192"
+        "src": "img/icons/icon-72x72.png",
+        "sizes": "72x72",
+        "type": "image/png"
       },
       {
-        "src": "/img/dota-2-logo-512.png",
-        "type": "image/png",
-        "sizes": "512x512"
+        "src": "img/icons/icon-96x96.png",
+        "sizes": "95x96",
+        "type": "image/png"
+      },
+      {
+        "src": "img/icons/icon-128x128.png",
+        "sizes": "127x128",
+        "type": "image/png"
+      },
+      {
+        "src": "img/icons/icon-144x144.png",
+        "sizes": "144x144",
+        "type": "image/png"
+      },
+      {
+        "src": "img/icons/icon-152x152.png",
+        "sizes": "151x152",
+        "type": "image/png"
+      },
+      {
+        "src": "img/icons/icon-192x192.png",
+        "sizes": "190x192",
+        "type": "image/png"
+      },
+      {
+        "src": "img/icons/icon-384x384.png",
+        "sizes": "380x384",
+        "type": "image/png"
+      },
+      {
+        "src": "img/icons/icon-512x512.png",
+        "sizes": "396x401",
+        "type": "image/png"
       }
     ],
-    "start_url": "/?source=pwa",
+    "start_url": "/",
     "background_color": "rgb(39, 39, 39)",
     "display": "standalone",
     "scope": "/",
