+++ docs/views/pages/detail.ejs
@@ -17,4 +17,5 @@
         <p></p>
         <p></p>
     </main>
-</body>
+
+    <%- include('../partials/pageend.ejs') %>
