+++ docs/views/pages/heroes.ejs
@@ -16,4 +16,5 @@
             <% }) %>
         </ul>
     </main>
-</body>
+
+    <%- include('../partials/pageend.ejs') %>
