+++ docs/views/pages/home.ejs
@@ -1,4 +1,6 @@
 <%- include('../partials/head.ejs') %>
 <body>
     <%- include('../partials/header.ejs') %>
-</body>
+
+
+    <%- include('../partials/pageend.ejs') %>
