+++ docs/views/pages/myStats.ejs
@@ -16,4 +16,4 @@
     </aside>
     <%- include('../partials/recentMatches.ejs') %>
 
-</body>
+<%- include('../partials/pageend.ejs') %>
