+++ docs/views/partials/head.ejs
@@ -6,6 +6,6 @@
     <link rel="stylesheet" href="/css/index.css">
     <link rel="manifest" href="/manifest.webmanifest">
     <link rel="shortcut icon" href="/img/favicon.ico" type="image/x-icon">
-    <script defer src="/js/script.js"></script>
+    
     <title><%= title %></title>
 </head>
