+++ docs/views/partials/pageend.ejs
@@ -1,0 +1,15 @@
+<script>
+    if ('serviceWorker' in navigator) {
+window.addEventListener('load', function() {
+  navigator.serviceWorker.register('/sw.js').then(function(registration) {
+    // Registration was successful
+    console.log('ServiceWorker registration successful with scope: ', registration.scope);
+    return registration.update();
+  }, function(err) {
+    // registration failed :(
+    console.log('ServiceWorker registration failed: ', err);
+  });
+});
+}
+</script>
+</body>
